pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arxiv-translator"
version = "0.3.0"
description = "Translates arXiv LaTeX papers to Chinese using Google Gemini models."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = ["arxiv", "translator", "gemini", "latex", "ai"]
authors = [
  { name = "Zeyu Chen", email = "zeyu@example.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "google-genai>=0.1.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
arxiv-translator = "arxiv_translator.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/arxiv_translator"]
